refactor(samples): create multiple microservices from existing monolith
