+++ SimpleEventSourcing.Samples.Events/AppEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace SimpleEventSourcing.Samples.Events
+{
+    public class AppEvent : SimpleEvent
+    {
+        public string Id { get; set; }
+    }
+}
