+++ SimpleEventSourcing.Samples.History/AppEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace SimpleEventSourcing.Samples.History
+{
+    public class AppEvent : SimpleEvent
+    {
+        public string Id { get; set; }
+    }
+}
