--- SimpleEventSourcing.Samples.Web/AppEvent.cs
@@ -1,7 +1,0 @@
-﻿namespace SimpleEventSourcing.Samples.Web
-{
-    public class AppEvent : SimpleEvent
-    {
-        public int? Id { get; set; }
-    }
-}
