--- SimpleEventSourcing.Samples.Web/Queries.cs
@@ -1,10 +1,0 @@
-﻿namespace SimpleEventSourcing.Samples.Web
-{
-    public class GetCartQuery { }
-
-    public class GetItemsQuery { }
-
-    public class GetOrdersQuery { }
-
-    public class GetEventsQuery { }
-}
