+++ SimpleEventSourcing.sln
@@ -1,13 +1,27 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2035
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28407.52
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing", "SimpleEventSourcing\SimpleEventSourcing.csproj", "{79FF39A6-846F-43D5-A4C4-6FC6F556D7E2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.UnitTests", "SimpleEventSourcing.UnitTests\SimpleEventSourcing.UnitTests.csproj", "{B2C658DC-9B92-4B75-8BBE-3863E5DECCFA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.Samples.Web", "SimpleEventSourcing.Samples.Web\SimpleEventSourcing.Samples.Web.csproj", "{BCACC13B-108C-4BAB-BB8C-82961990A34E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{B19CB91A-294A-4D23-BAC6-3E5B1E2AB125}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{A78DC66A-BCED-4E06-B589-52358366FD92}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.Samples.Shop", "SimpleEventSourcing.Samples.Shop\SimpleEventSourcing.Samples.Shop.csproj", "{1CAA1F7D-7F4E-4B9A-8DD0-F1130694BB59}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.Samples.Inventory", "SimpleEventSourcing.Samples.Inventory\SimpleEventSourcing.Samples.Inventory.csproj", "{FB6806AD-F18C-4278-B1A3-484889E10393}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.Samples.Orders", "SimpleEventSourcing.Samples.Orders\SimpleEventSourcing.Samples.Orders.csproj", "{DEEEA1A3-8F0B-4791-8B6C-E84125177959}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.Samples.Events", "SimpleEventSourcing.Samples.Events\SimpleEventSourcing.Samples.Events.csproj", "{E08EEF9B-7EDE-49D8-97B1-FC0833B61874}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleEventSourcing.Samples.History", "SimpleEventSourcing.Samples.History\SimpleEventSourcing.Samples.History.csproj", "{6A9C5D41-01A0-4BA9-A925-B2509FF8FAF9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,9 +41,39 @@ Global
 		{BCACC13B-108C-4BAB-BB8C-82961990A34E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BCACC13B-108C-4BAB-BB8C-82961990A34E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BCACC13B-108C-4BAB-BB8C-82961990A34E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1CAA1F7D-7F4E-4B9A-8DD0-F1130694BB59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1CAA1F7D-7F4E-4B9A-8DD0-F1130694BB59}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1CAA1F7D-7F4E-4B9A-8DD0-F1130694BB59}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1CAA1F7D-7F4E-4B9A-8DD0-F1130694BB59}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FB6806AD-F18C-4278-B1A3-484889E10393}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB6806AD-F18C-4278-B1A3-484889E10393}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB6806AD-F18C-4278-B1A3-484889E10393}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB6806AD-F18C-4278-B1A3-484889E10393}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DEEEA1A3-8F0B-4791-8B6C-E84125177959}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DEEEA1A3-8F0B-4791-8B6C-E84125177959}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DEEEA1A3-8F0B-4791-8B6C-E84125177959}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DEEEA1A3-8F0B-4791-8B6C-E84125177959}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E08EEF9B-7EDE-49D8-97B1-FC0833B61874}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E08EEF9B-7EDE-49D8-97B1-FC0833B61874}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E08EEF9B-7EDE-49D8-97B1-FC0833B61874}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E08EEF9B-7EDE-49D8-97B1-FC0833B61874}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6A9C5D41-01A0-4BA9-A925-B2509FF8FAF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A9C5D41-01A0-4BA9-A925-B2509FF8FAF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A9C5D41-01A0-4BA9-A925-B2509FF8FAF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A9C5D41-01A0-4BA9-A925-B2509FF8FAF9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{79FF39A6-846F-43D5-A4C4-6FC6F556D7E2} = {B19CB91A-294A-4D23-BAC6-3E5B1E2AB125}
+		{B2C658DC-9B92-4B75-8BBE-3863E5DECCFA} = {B19CB91A-294A-4D23-BAC6-3E5B1E2AB125}
+		{BCACC13B-108C-4BAB-BB8C-82961990A34E} = {A78DC66A-BCED-4E06-B589-52358366FD92}
+		{1CAA1F7D-7F4E-4B9A-8DD0-F1130694BB59} = {A78DC66A-BCED-4E06-B589-52358366FD92}
+		{FB6806AD-F18C-4278-B1A3-484889E10393} = {A78DC66A-BCED-4E06-B589-52358366FD92}
+		{DEEEA1A3-8F0B-4791-8B6C-E84125177959} = {A78DC66A-BCED-4E06-B589-52358366FD92}
+		{E08EEF9B-7EDE-49D8-97B1-FC0833B61874} = {A78DC66A-BCED-4E06-B589-52358366FD92}
+		{6A9C5D41-01A0-4BA9-A925-B2509FF8FAF9} = {A78DC66A-BCED-4E06-B589-52358366FD92}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {23EAD2F7-63C4-4362-8C89-C93E92CEE47A}
